Добавление Realm через Swift Package Dependency Удобнее и красивее (говорят скорость больше) но на рабочем проекте выеживался Попробовать с нуля в следующий раз ))
