--- Lesson20/Model/Task.swift
@@ -1,11 +1,0 @@
-import RealmSwift
-
-//Так как наследуемся от другого класса Object (typealias Object = RealmSwiftObject)
-//Что бы молги определить свои модели как обычные классы Swift
-class Task: Object
-{
-    @objc dynamic var name = ""
-    @objc dynamic var note = ""
-    @objc dynamic var date = Data()
-    @objc dynamic var isComplete = false
-}
