+++ Lesson20/System32/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -2,6 +2,18 @@
   "images" : [
     {
       "size" : "20x20",
+      "idiom" : "ipad",
+      "filename" : "IconSmall-20.png",
+      "scale" : "1x"
+    },
+    {
+      "size" : "20x20",
+      "idiom" : "ipad",
+      "filename" : "IconSmall-20-iPad@2x.png",
+      "scale" : "2x"
+    },
+    {
+      "size" : "20x20",
       "idiom" : "iphone",
       "filename" : "IconSmall-20@2x.png",
       "scale" : "2x"
@@ -47,6 +59,48 @@
       "idiom" : "iphone",
       "filename" : "Icon@3x.png",
       "scale" : "3x"
+    },
+    {
+      "size" : "29x29",
+      "idiom" : "ipad",
+      "filename" : "IconSmall.png",
+      "scale" : "1x"
+    },
+    {
+      "size" : "29x29",
+      "idiom" : "ipad",
+      "filename" : "IconSmall-iPad@2x.png",
+      "scale" : "2x"
+    },
+    {
+      "size" : "40x40",
+      "idiom" : "ipad",
+      "filename" : "IconSmall-40-iPad@2x.png",
+      "scale" : "2x"
+    },
+    {
+      "size" : "40x40",
+      "idiom" : "ipad",
+      "filename" : "IconSmall-40.png",
+      "scale" : "1x"
+    },
+    {
+      "size" : "76x76",
+      "idiom" : "ipad",
+      "filename" : "Icon-76.png",
+      "scale" : "1x"
+    },
+    {
+      "size" : "76x76",
+      "idiom" : "ipad",
+      "filename" : "Icon-76@2x.png",
+      "scale" : "2x"
+    },
+    {
+      "size" : "83.5x83.5",
+      "idiom" : "ipad",
+      "filename" : "Icon-83.5@2x.png",
+      "scale" : "2x"
     },
     {
       "size" : "1024x1024",
